+++ lib/dolitypology_typology.lib.php
@@ -41,6 +41,11 @@ function typologyPrepareHead($object)
 	$head[$h][2] = 'card';
 	$h++;
 
+	$head[$h][0] = dol_buildpath("/dolitypology/typology_relatedobjects.php", 1).'?id='.$object->id;
+	$head[$h][1] = $langs->trans("RelatedObjects").' <span class="badge">xx</span>';
+	$head[$h][2] = 'relatedobjects';
+	$h++;
+
 	if (isset($object->fields['note_public']) || isset($object->fields['note_private'])) {
 		$nbNote = 0;
 		if (!empty($object->note_private)) {