+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # CHANGELOG DOLITYPOLOGY FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
-## 1.0
+## 1.0.1
+- MAJ: Update module icon
+- NEW: Add fr_FR translation file
+- NEW: TypologyExtrafields to extends extrafields filters
+- NEW: Add a new tab for related objects
+- MAJ: Update extrafields list on typology card
 
+## 1.0.0
 Initial version